fixed a very silly bug
I had misplaced the quotes and commas, and didn't notice it
visually.

I was pretty sure the bug was in the called function, not in
the way I was calling it, because I've become very reliant
on TS's type checking.

But in this case TS saw nothing:

The function was expecting an array of strings and I was passing it
an array of strings! damn it! Took me ages to notice, even
with the debugger activated.