trying to fix hot reload - or did I?
I think that for "normal" things, for HMR to work,
I need to re-require the modules within the module.hot.accept
callback.

Did it for the reducers and for my custom server payload
handling stuff.

Hot-reloading of React is already working well,
it didn't enter the accept's callback,
I think it was working because of @pmmmwh/react-refresh-webpack-plugin
kind of automagically.