choose correct protocol for websockets
Using HTTPS happens when the node server
is behind apache and apache serves the site
over HTTPS.

The node app is assumed to be on the
same safe network as the apache server so
we don't bother handling SSL in node.

Tnere will always be an SSL endpoint in front of
this node server.